service vehicule, velo et voiture

+++ src/main/java/com/example/gestion_location_vehicule/model/Vehicule.java
@@ -25,7 +25,7 @@ public abstract class Vehicule {
 
     private String couleur;
 
-    private int notevehicule;
+    private double notevehicule;
 
     private boolean vehiculedispo;
